Refactored input logic for traversal.
The input for the directory traversal is done through an interface. This
will be useful in the future when this tools is made bash pipeable and
also for reporting the reporting in the input directory tree. Previously
the FileWalker was aware of the absolute paths in the input directory
which made it more cumbersome to generate the name of the omitted files
without the base input path.