Kubernetes Resource omitter
The resource omitter reads the contents of a json or yaml file and
tries to decode them as a kubernetes resource or a resource list. For a
resource list if the input criteria of kind, api version and namespaces
match then the omitter indicates this. If for a resource list one of the
item matches the criteria then the entire file is omitted.