Refactored the omit configuration schema
Currently the omit coniguration property can have multiple object types
as a value. As a result the type of the omit property is []interface{}
which makes validation and initialization of the omitters more
difficult. This change refactor the omit type to use a single type with
multiple configuration fields which can be validated in the application.