Concurrent processing of input directory
This change modifies the FileWalker to use workers which can process
files concurrently. Files are passed to the workers through channels and
an error channel returns any errors which might have been encountered
when processing a file. When the error channel returns and error the
FileWalker immediately exits with the received error.